requirements.txt
fastapi>=0.103.1
uvicorn>=0.23.2
openai>=1.3.0
python-dotenv>=1.0.0
pydantic>=2.4.2
loguru>=0.7.0 